+++ ui/feature/quizstartscreen/src/main/java/com/serelik/surfdailyquiz/quizstartscreen/QuizState.kt
@@ -4,5 +4,5 @@ sealed interface QuizState {
     data object NotStarted : QuizState
     data object Loading : QuizState
     data object Error : QuizState
-    data class Result(val questionIdList: List<Int>) : QuizState
+    data class Result(val questionIdList: List<Long>) : QuizState
 }
